refactor(frontend): redesign pending messages as queue cards above input bar
Move queued messages from the message list into a card anchored above the
input bar, matching its color scheme. Replace PendingMessage bubble with a
compact QueueMessageCard row layout with a CornerDownRight icon indicator.
